ShoppingProgram/requirements.txt
mysql-connector-python
requests
beautifulsoup4
lxml
python-dotenv
